requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
python-decouple==3.8
django-cors-headers==4.3.0
reportlab
qrcode
Pillow
openpyxl
django-filter==23.5
twilio>=9.0.0
django-apscheduler
groq
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
django-redis>=5.4.0
drf-spectacular>=0.27.0
